# _uvoz/bookie-uvoz-prejemnic/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ERP — Modul prevzema blaga
  description: |
    Uvoz prejemnic in dobavnic dobaviteljev za POS gostinstvo.
    Vsi kanali uvoza (datoteka, e-pošta, Peppol, API, OCR) se stekajo
    v isti tok obdelave: zajem → razčlenitev → uparjanje → validacija → knjiženje.
  version: "1.0.0"

# ⚠ OPOMBA: ta datoteka je bila pisana kot pogodba pred izvedbo.
# Dejanske poti so /api/pos/uvoz/... in so definirane v
# artifacts/api-server/src/routes/pos/uvoz-prejemnic.ts in uvoz-profili.ts.
# Ob razhajanju velja koda. Datoteka ostaja kot pregled obsega vmesnika.
servers:
  - url: /api/pos/uvoz
security:
  - bearerAuth: []
...
paths:
...
  /uvoz/datoteka:
    post:
      tags: [Uvoz]
      summary: Naloži dokument za uvoz
      description: |
        Sprejme datoteko poljubnega podprtega formata. Format se zazna
        avtomatsko, razen če je podan izrecno. Vrne ID uvozne seje.
        Če je bila enaka datoteka (SHA-256) že obdelana, vrne 409 s
        sklicem na obstoječo sejo.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [datoteka]
              properties:
                datoteka:
                  type: string
                  format: binary
                poslovalnica_id:
                  type: integer
                  format: int64
                dobavitelj_id:
                  type: integer
                  format: int64
                  description: Neobvezno; pospeši izbiro uvoznega profila.
                format:
                  $ref: '#/components/schemas/UvozFormat'
                takoj_obdelaj:
                  type: boolean
                  default: true
      responses:
        '202':
          description: Sprejeto v obdelavo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UvozSeja'
        '409':
          description: Podvojen dokument
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Napaka'
                  - type: object
                    properties:
                      obstojeca_seja_id: { type: string, format: uuid }
                      obstojeci_prejem_id: { type: integer, format: int64 }
        '415': { $ref: '#/components/responses/NepodprtFormat' }
  /uvoz/seje:
    get:
      tags: [Uvoz]
      summary: Seznam uvoznih sej
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/UvozStatus' } }
        - { name: kanal, in: query, schema: { $ref: '#/components/schemas/UvozKanal' } }
        - { name: od, in: query, schema: { type: string, format: date } }
        - { name: do, in: query, schema: { type: string, format: date } }
        - { $ref: '#/components/parameters/Stran' }
        - { $ref: '#/components/parameters/Velikost' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  vsebina:
                    type: array
                    items: { $ref: '#/components/schemas/UvozSeja' }
                  skupaj: { type: integer }
...
  /uvoz/seje/{sejaId}/izvirnik:
    get:
      tags: [Uvoz]
      summary: Prenos izvorne datoteke
      description: Vrne nespremenjeno vhodno datoteko za revizijske namene.
      parameters: [{ $ref: '#/components/parameters/SejaId' }]
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
  /uvoz/seje/{sejaId}/ponovi:
    post:
      tags: [Uvoz]
      summary: Ponovna obdelava iz izvirnika
      description: |
        Uporabi po popravku uvoznega profila ali preslikav artiklov.
        Ustvari novo sejo s sklicem na prvotno.
      parameters: [{ $ref: '#/components/parameters/SejaId' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                profil_id: { type: integer, format: int64 }
                format: { $ref: '#/components/schemas/UvozFormat' }
      responses:
        '202':
          description: Sprejeto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UvozSeja' }
  /uvoz/predogled:
    post:
      tags: [Uvoz]
      summary: Razčleni brez shranjevanja
      description: |
        Za nastavljanje uvoznega profila: vrne razčlenjen rezultat,
        ne ustvari seje ne prejemnice.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                datoteka: { type: string, format: binary }
                konfiguracija:
                  type: string
                  description: JSON konfiguracije profila za preizkus
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  zaznan_format: { $ref: '#/components/schemas/UvozFormat' }
                  dto: { $ref: '#/components/schemas/PrejemDTO' }
                  napake:
                    type: array
                    items: { $ref: '#/components/schemas/Opozorilo' }

  # ==================== PREJEMNICE ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SejaId:
      name: sejaId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Stran:
      name: stran
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
    Velikost:
      name: velikost
      in: query
      schema: { type: integer, default: 50, maximum: 500 }
  responses:
...
    NepodprtFormat:
      description: Format datoteke ni podprt
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Napaka' }
  schemas:
    UvozKanal:
      type: string
      enum: [ROCNI_NALOG, NADZOROVANA_MAPA, EPOSTA_IMAP, EPOT_PONUDNIK,
             PEPPOL_AP, API_DOBAVITELJ, FOTOAPARAT, CRTNA_KODA,
             ROCNI_VNOS, IZ_NAROCILNICE]
    UvozFormat:
      type: string
      enum: [ESLOG_2_0_DOBAVNICA, ESLOG_2_0_RACUN, ESLOG_1_6_1,
             UBL_2_1_DESPATCH, UBL_2_1_INVOICE, CII_D16B,
             EDIFACT_DESADV, EDIFACT_INVOIC, EDIFACT_PRICAT, GS1_XML,
             CSV, FIKSNA_SIRINA, XLSX, JSON_LASTNI, XML_LASTNI,
             PDF_PREDLOGA, PDF_OCR, SLIKA_OCR, BREZ]
    UvozStatus:
      type: string
      enum: [PREJETO, RAZCLENJENO, NAPAKA_RAZCLENITVE, PODVOJENO, OBDELANO, ZAVRZENO]
...
    UvozSeja:
      type: object
      properties:
        id: { type: string, format: uuid }
        kanal: { $ref: '#/components/schemas/UvozKanal' }
        format: { $ref: '#/components/schemas/UvozFormat' }
        ime_datoteke: { type: string }
        velikost_b: { type: integer, format: int64 }
        sha256: { type: string }
        prejeto: { type: string, format: date-time }
        obdelano: { type: string, format: date-time, nullable: true }
        status: { $ref: '#/components/schemas/UvozStatus' }
        prejem_id: { type: integer, format: int64, nullable: true }
        napake:
          type: array
          items: { $ref: '#/components/schemas/Opozorilo' }
    PrejemDTO:
      description: Kanonični vmesni model, enak ne glede na izvorni format.
      type: object
      properties:
        dobavitelj:
          type: object
          properties:
            davcna: { type: string }
            maticna: { type: string }
            gln: { type: string }
            naziv: { type: string }
            drzava: { type: string }
        dokument:
          type: object
          properties:
            vrsta: { type: string, enum: [DOBAVNICA, RACUN, DOBROPIS, BREMEPIS] }
            stevilka: { type: string }
            datum_izdaje: { type: string, format: date }
            datum_dobave: { type: string, format: date }
            sklic_narocilnica: { type: string }
            valuta: { type: string }
            tecaj: { type: number }
        postavke:
          type: array
          items: { $ref: '#/components/schemas/PrejemDTOPostavka' }
        odvisni_stroski:
          type: array
          items:
            type: object
            properties:
              vrsta: { type: string, enum: [PREVOZ, CARINA, MANIPULACIJA, ZAVAROVANJE, DRUGO] }
              znesek: { type: number }
              ddv_stopnja: { type: number }
        povratna_embalaza:
          type: array
          items:
            type: object
            properties:
              gtin: { type: string }
              kolicina: { type: number }
              kavcija: { type: number }
        vsote:
          type: object
          properties:
            osnova: { type: number }
            ddv: { type: number }
            skupaj: { type: number }
    PrejemDTOPostavka:
      type: object
      properties:
        zap: { type: integer }
        gtin: { type: string, nullable: true }
        sifra_dobavitelja: { type: string, nullable: true }
        naziv: { type: string }
        kolicina: { type: number }
        enota: { type: string, description: UN/ECE Rec 20 ali interna koda }
        cena_bruto: { type: number, nullable: true }
        cena_neto: { type: number, nullable: true }
        cena_osnova_kol: { type: number, default: 1 }
        rabat_odstotek: { type: number, nullable: true }
        ddv_stopnja: { type: number, nullable: true }
        ddv_kategorija: { type: string, description: 'UNCL5305: S, Z, E, AE, K, G, O' }
        vrednost_neto: { type: number, nullable: true }
        lot: { type: string, nullable: true }
        rok_uporabe: { type: string, format: date, nullable: true }
        sscc: { type: string, nullable: true }
        zaupanje:
          type: number
          description: 0..1, relevantno pri OCR/LLM viru
...
    Opozorilo:
      type: object
      properties:
        koda: { type: string, example: POS006 }
        resnost: { type: string, enum: [B, O, I] }
        sporocilo: { type: string }
        polje: { type: string, nullable: true }
        postavka_id: { type: integer, format: int64, nullable: true }
        podatki: { type: object, additionalProperties: true }
        potrdil_uporabnik: { type: integer, format: int64, nullable: true }
        potrjeno_dne: { type: string, format: date-time, nullable: true }
        opomba: { type: string, nullable: true }
    Napaka:
      type: object
      properties:
        koda: { type: string }
        sporocilo: { type: string }
        podrobnosti: { type: object, additionalProperties: true }
        cas: { type: string, format: date-time }